centering around two classes with BaseIntegrator
